Create application grant selector component
